+++ HISTORY.rst
@@ -1,5 +1,15 @@
 Bug tracker at https://github.com/giampaolo/pyftpdlib/issues
 
+Version: 1.6.0 - XXXX-XX-XX
+===========================
+
+**Enhancements**
+
+- #378: SSL security was massively improved disabling insecure feautures such
+  as SSLv2, SSLv3 and SSL_COMPRESSION. New TLS_FTPHandler's ssl_options
+  and ssl_ciphers class attributes were added.
+
+
 Version: 1.5.0 - 2015-12-13
 ===========================